+++ familyhelper-webapi-node/src/main/java/com/dwarfeng/familyhelper/webapi/node/controller/v1/project/MemoFileController.java
@@ -1,0 +1,342 @@
+package com.dwarfeng.familyhelper.webapi.node.controller.v1.project;
+
+import com.dwarfeng.dutil.basic.io.IOUtil;
+import com.dwarfeng.familyhelper.project.sdk.bean.entity.JSFixedFastJsonMemoFileInfo;
+import com.dwarfeng.familyhelper.project.stack.bean.dto.MemoFile;
+import com.dwarfeng.familyhelper.project.stack.bean.dto.MemoFileUpdateInfo;
+import com.dwarfeng.familyhelper.project.stack.bean.dto.MemoFileUploadInfo;
+import com.dwarfeng.familyhelper.project.stack.bean.entity.MemoFileInfo;
+import com.dwarfeng.familyhelper.webapi.stack.handler.system.TokenHandler;
+import com.dwarfeng.familyhelper.webapi.stack.service.project.MemoFileResponseService;
+import com.dwarfeng.subgrade.sdk.bean.dto.FastJsonResponseData;
+import com.dwarfeng.subgrade.sdk.bean.dto.JSFixedFastJsonPagedData;
+import com.dwarfeng.subgrade.sdk.bean.dto.PagingUtil;
+import com.dwarfeng.subgrade.sdk.bean.dto.ResponseDataUtil;
+import com.dwarfeng.subgrade.sdk.bean.key.WebInputLongIdKey;
+import com.dwarfeng.subgrade.sdk.interceptor.analyse.BehaviorAnalyse;
+import com.dwarfeng.subgrade.sdk.interceptor.http.BindingCheck;
+import com.dwarfeng.subgrade.sdk.interceptor.login.LoginRequired;
+import com.dwarfeng.subgrade.stack.bean.BeanTransformer;
+import com.dwarfeng.subgrade.stack.bean.dto.PagedData;
+import com.dwarfeng.subgrade.stack.bean.dto.PagingInfo;
+import com.dwarfeng.subgrade.stack.bean.key.LongIdKey;
+import com.dwarfeng.subgrade.stack.bean.key.StringIdKey;
+import com.dwarfeng.subgrade.stack.exception.ServiceExceptionMapper;
+import org.springframework.http.HttpHeaders;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.*;
+import org.springframework.web.multipart.MultipartFile;
+import org.springframework.web.multipart.MultipartHttpServletRequest;
+import org.springframework.web.multipart.commons.CommonsMultipartResolver;
+
+import javax.servlet.http.HttpServletRequest;
+import java.io.ByteArrayOutputStream;
+import java.io.InputStream;
+import java.nio.charset.StandardCharsets;
+import java.util.Objects;
+
+/**
+ * 备忘录文件控制器。
+ *
+ * @author DwArFeng
+ * @since 1.0.3
+ */
+@RestController
+@RequestMapping("/api/v1/project")
+public class MemoFileController {
+
+    /**
+     * IO 传输设定的缓冲容量。
+     */
+    private static final int IO_TRANS_BUFFER_SIZE = 4096;
+
+    private final MemoFileResponseService service;
+
+    private final ServiceExceptionMapper sem;
+
+    private final BeanTransformer<MemoFileInfo, JSFixedFastJsonMemoFileInfo> beanTransformer;
+
+    private final CommonsMultipartResolver commonsMultipartResolver;
+
+    private final TokenHandler tokenHandler;
+
+    public MemoFileController(
+            MemoFileResponseService service,
+            ServiceExceptionMapper sem,
+            BeanTransformer<MemoFileInfo, JSFixedFastJsonMemoFileInfo> beanTransformer,
+            CommonsMultipartResolver commonsMultipartResolver,
+            TokenHandler tokenHandler
+    ) {
+        this.service = service;
+        this.sem = sem;
+        this.beanTransformer = beanTransformer;
+        this.commonsMultipartResolver = commonsMultipartResolver;
+        this.tokenHandler = tokenHandler;
+    }
+
+    @GetMapping("/memo-file/{longId}/exists")
+    @BehaviorAnalyse
+    @LoginRequired
+    public FastJsonResponseData<Boolean> exists(HttpServletRequest request, @PathVariable("longId") Long longId) {
+        try {
+            boolean exists = service.exists(new LongIdKey(longId));
+            return FastJsonResponseData.of(ResponseDataUtil.good(exists));
+        } catch (Exception e) {
+            return FastJsonResponseData.of(ResponseDataUtil.bad(Boolean.class, e, sem));
+        }
+    }
+
+    @GetMapping("/memo-file/{longId}")
+    @BehaviorAnalyse
+    @LoginRequired
+    public FastJsonResponseData<JSFixedFastJsonMemoFileInfo> get(
+            HttpServletRequest request, @PathVariable("longId") Long longId
+    ) {
+        try {
+            MemoFileInfo memoFileInfo = service.get(new LongIdKey(longId));
+            return FastJsonResponseData.of(ResponseDataUtil.good(JSFixedFastJsonMemoFileInfo.of(memoFileInfo)));
+        } catch (Exception e) {
+            return FastJsonResponseData.of(ResponseDataUtil.bad(JSFixedFastJsonMemoFileInfo.class, e, sem));
+        }
+    }
+
+    @GetMapping("memo/{memoId}/memo-file/default")
+    @BehaviorAnalyse
+    @LoginRequired
+    public FastJsonResponseData<JSFixedFastJsonPagedData<JSFixedFastJsonMemoFileInfo>> childForMemo(
+            HttpServletRequest request,
+            @PathVariable("memoId") long memoId,
+            @RequestParam("page") int page, @RequestParam("rows") int rows
+    ) {
+        try {
+            PagedData<MemoFileInfo> childForMemo = service.childForMemo(
+                    new LongIdKey(memoId), new PagingInfo(page, rows)
+            );
+            PagedData<JSFixedFastJsonMemoFileInfo> transform = PagingUtil.transform(
+                    childForMemo, beanTransformer
+            );
+            return FastJsonResponseData.of(ResponseDataUtil.good(JSFixedFastJsonPagedData.of(transform)));
+        } catch (Exception e) {
+            e.printStackTrace();
+            return FastJsonResponseData.of(ResponseDataUtil.bad(JSFixedFastJsonPagedData.class, e, sem));
+        }
+    }
+
+    @GetMapping("memo/{memoId}/memo-file/inspected-date-desc")
+    @BehaviorAnalyse
+    @LoginRequired
+    public FastJsonResponseData<JSFixedFastJsonPagedData<JSFixedFastJsonMemoFileInfo>> childForMemoInspectedDateDesc(
+            HttpServletRequest request,
+            @PathVariable("memoId") long memoId,
+            @RequestParam("page") int page, @RequestParam("rows") int rows
+    ) {
+        try {
+            PagedData<MemoFileInfo> childForMemo = service.childForMemoInspectedDateDesc(
+                    new LongIdKey(memoId), new PagingInfo(page, rows)
+            );
+            PagedData<JSFixedFastJsonMemoFileInfo> transform = PagingUtil.transform(
+                    childForMemo, beanTransformer
+            );
+            return FastJsonResponseData.of(ResponseDataUtil.good(JSFixedFastJsonPagedData.of(transform)));
+        } catch (Exception e) {
+            return FastJsonResponseData.of(ResponseDataUtil.bad(JSFixedFastJsonPagedData.class, e, sem));
+        }
+    }
+
+    @GetMapping("memo/{memoId}/memo-file/modified-date-desc")
+    @BehaviorAnalyse
+    @LoginRequired
+    public FastJsonResponseData<JSFixedFastJsonPagedData<JSFixedFastJsonMemoFileInfo>> childForMemoModifiedDateDesc(
+            HttpServletRequest request,
+            @PathVariable("memoId") long memoId,
+            @RequestParam("page") int page, @RequestParam("rows") int rows
+    ) {
+        try {
+            PagedData<MemoFileInfo> childForMemo = service.childForMemoModifiedDateDesc(
+                    new LongIdKey(memoId), new PagingInfo(page, rows)
+            );
+            PagedData<JSFixedFastJsonMemoFileInfo> transform = PagingUtil.transform(
+                    childForMemo, beanTransformer
+            );
+            return FastJsonResponseData.of(ResponseDataUtil.good(JSFixedFastJsonPagedData.of(transform)));
+        } catch (Exception e) {
+            return FastJsonResponseData.of(ResponseDataUtil.bad(JSFixedFastJsonPagedData.class, e, sem));
+        }
+    }
+
+    @GetMapping("memo/{memoId}/memo-file/origin-name-asc")
+    @BehaviorAnalyse
+    @LoginRequired
+    public FastJsonResponseData<JSFixedFastJsonPagedData<JSFixedFastJsonMemoFileInfo>> childForMemoOriginNameAsc(
+            HttpServletRequest request,
+            @PathVariable("memoId") long memoId,
+            @RequestParam("page") int page, @RequestParam("rows") int rows
+    ) {
+        try {
+            PagedData<MemoFileInfo> childForMemo = service.childForMemoOriginNameAsc(
+                    new LongIdKey(memoId), new PagingInfo(page, rows)
+            );
+            PagedData<JSFixedFastJsonMemoFileInfo> transform = PagingUtil.transform(
+                    childForMemo, beanTransformer
+            );
+            return FastJsonResponseData.of(ResponseDataUtil.good(JSFixedFastJsonPagedData.of(transform)));
+        } catch (Exception e) {
+            return FastJsonResponseData.of(ResponseDataUtil.bad(JSFixedFastJsonPagedData.class, e, sem));
+        }
+    }
+
+    @GetMapping("memo/{memoId}/memo-file/created-date-asc")
+    @BehaviorAnalyse
+    @LoginRequired
+    public FastJsonResponseData<JSFixedFastJsonPagedData<JSFixedFastJsonMemoFileInfo>> childForMemoCreatedDateAsc(
+            HttpServletRequest request,
+            @PathVariable("memoId") long memoId,
+            @RequestParam("page") int page, @RequestParam("rows") int rows
+    ) {
+        try {
+            PagedData<MemoFileInfo> childForMemo = service.childForMemoCreatedDateAsc(
+                    new LongIdKey(memoId), new PagingInfo(page, rows)
+            );
+            PagedData<JSFixedFastJsonMemoFileInfo> transform = PagingUtil.transform(
+                    childForMemo, beanTransformer
+            );
+            return FastJsonResponseData.of(ResponseDataUtil.good(JSFixedFastJsonPagedData.of(transform)));
+        } catch (Exception e) {
+            return FastJsonResponseData.of(ResponseDataUtil.bad(JSFixedFastJsonPagedData.class, e, sem));
+        }
+    }
+
+
+    @GetMapping("/memo-file/{memoFileId}/download")
+    @BehaviorAnalyse
+    @BindingCheck
+    @LoginRequired
+    public ResponseEntity<Object> downloadMemoFile(
+            HttpServletRequest request, @PathVariable("memoFileId") Long memoFileId
+    ) {
+        HttpHeaders headers = new HttpHeaders();
+        Object body;
+        try {
+            StringIdKey accountKey = tokenHandler.getAccountKey(request);
+            MemoFile memoFile = service.downloadMemoFile(accountKey, new LongIdKey(memoFileId));
+            // 将文件名转换成 HTTP 标准文件名编码下的格式。
+            String fileName = adjustFileNameEncoding(memoFile.getOriginName());
+            headers.add("Content-Disposition", "attachment;filename=" + fileName);
+            body = memoFile.getContent();
+        } catch (Exception e) {
+            body = FastJsonResponseData.of(ResponseDataUtil.bad(Object.class, e, sem));
+        }
+        return new ResponseEntity<>(body, headers, HttpStatus.OK);
+    }
+
+    @SuppressWarnings("DuplicatedCode")
+    @PostMapping("/memo/{memoId}/memo-file/upload")
+    @BehaviorAnalyse
+    @BindingCheck
+    @LoginRequired
+    public FastJsonResponseData<Object> uploadMemoFile(
+            HttpServletRequest request, @PathVariable("memoId") Long memoId
+    ) {
+        try {
+            // 通过请求解析用户。
+            StringIdKey accountKey = tokenHandler.getAccountKey(request);
+
+            // 确认请求合法。
+            if (!commonsMultipartResolver.isMultipart(request)) {
+                throw new IllegalStateException("请求不是标准的文件上传请求");
+            }
+
+            //获取 multiRequest 中的文件。
+            MultipartHttpServletRequest multipartHttpServletRequest = commonsMultipartResolver.resolveMultipart(request);
+            MultipartFile file = multipartHttpServletRequest.getFile("file");
+            if (Objects.isNull(file)) {
+                throw new IllegalStateException("请求体中缺少 file 属性");
+            }
+
+            // 解析文件内容。
+            String originFileName = file.getOriginalFilename();
+            byte[] content;
+            try (InputStream in = file.getInputStream(); ByteArrayOutputStream bout = new ByteArrayOutputStream()) {
+                IOUtil.trans(in, bout, IO_TRANS_BUFFER_SIZE);
+                bout.flush();
+                content = bout.toByteArray();
+            }
+
+            // 将文件内容转换为接口需要的格式，并上传。
+            service.uploadMemoFile(
+                    accountKey, new MemoFileUploadInfo(new LongIdKey(memoId), originFileName, content)
+            );
+
+            // 返回响应结果。
+            return FastJsonResponseData.of(ResponseDataUtil.good(null));
+        } catch (Exception e) {
+            return FastJsonResponseData.of(ResponseDataUtil.bad(Object.class, e, sem));
+        }
+    }
+
+    @SuppressWarnings("DuplicatedCode")
+    @PostMapping("/memo-file/{memoFileId}/update")
+    @BehaviorAnalyse
+    @BindingCheck
+    @LoginRequired
+    public FastJsonResponseData<Object> updateMemoFile(
+            HttpServletRequest request, @PathVariable("memoFileId") Long memoFileId
+    ) {
+        try {
+            // 通过请求解析用户。
+            StringIdKey accountKey = tokenHandler.getAccountKey(request);
+
+            // 确认请求合法。
+            if (!commonsMultipartResolver.isMultipart(request)) {
+                throw new IllegalStateException("请求不是标准的文件上传请求");
+            }
+
+            //获取 multiRequest 中的文件。
+            MultipartHttpServletRequest multipartHttpServletRequest = commonsMultipartResolver.resolveMultipart(request);
+            MultipartFile file = multipartHttpServletRequest.getFile("file");
+            if (Objects.isNull(file)) {
+                throw new IllegalStateException("请求体中缺少 file 属性");
+            }
+
+            // 解析文件内容。
+            String originFileName = file.getOriginalFilename();
+            byte[] content;
+            try (InputStream in = file.getInputStream(); ByteArrayOutputStream bout = new ByteArrayOutputStream()) {
+                IOUtil.trans(in, bout, IO_TRANS_BUFFER_SIZE);
+                bout.flush();
+                content = bout.toByteArray();
+            }
+
+            // 将文件内容转换为接口需要的格式，并上传。
+            service.updateMemoFile(
+                    accountKey, new MemoFileUpdateInfo(new LongIdKey(memoFileId), originFileName, content)
+            );
+
+            // 返回响应结果。
+            return FastJsonResponseData.of(ResponseDataUtil.good(null));
+        } catch (Exception e) {
+            return FastJsonResponseData.of(ResponseDataUtil.bad(Object.class, e, sem));
+        }
+    }
+
+    @PostMapping("/memo-file/remove")
+    @BehaviorAnalyse
+    @BindingCheck
+    @LoginRequired
+    public FastJsonResponseData<Object> removeMemoFile(
+            HttpServletRequest request, @RequestBody WebInputLongIdKey memoFileKey
+    ) {
+        try {
+            StringIdKey accountKey = tokenHandler.getAccountKey(request);
+            service.removeMemoFile(accountKey, WebInputLongIdKey.toStackBean(memoFileKey));
+            return FastJsonResponseData.of(ResponseDataUtil.good(null));
+        } catch (Exception e) {
+            return FastJsonResponseData.of(ResponseDataUtil.bad(Object.class, e, sem));
+        }
+    }
+
+    private String adjustFileNameEncoding(String fileName) {
+        return new String(fileName.getBytes(), StandardCharsets.ISO_8859_1);
+    }
+}